member 내부에 auth 위치하도록 패키지 구조 수정
