+++ src/main/java/com/example/eshop/member/auth/repository/AuthRepository.java
@@ -1,16 +1,12 @@
-package com.example.eshop.auth.repository;
+package com.example.eshop.member.auth.repository;
 
-import com.example.eshop.auth.model.TokenEntity;
-import com.example.eshop.auth.model.UserEntity;
+import com.example.eshop.member.auth.model.TokenEntity;
 import org.apache.ibatis.annotations.Mapper;
 import org.springframework.stereotype.Repository;
 
 @Mapper
 @Repository
 public interface AuthRepository {
-    boolean isDuplicatedId(String userId);
-    void signin(UserEntity user);
-    UserEntity findUserByUserId(String userId);
     TokenEntity findAccessTokenByUserNo(long userNo);
     TokenEntity findAccessTokenByRandomToken(String randomToken);
     TokenEntity findRefreshTokenByRandomToken(String randomToken);
