+++ src/main/java/com/example/eshop/member/auth/service/AuthService.java
@@ -1,12 +1,9 @@
-package com.example.eshop.auth.service;
+package com.example.eshop.member.auth.service;
 
 import com.example.eshop.controller.dto.LoginDto;
 import com.example.eshop.controller.dto.TokenDto;
-import com.example.eshop.controller.dto.UserDto;
 
 public interface AuthService {
-    boolean isDuplicatedId(String id);
-    void signin(UserDto userDto);
     TokenDto login(LoginDto loginDto);
     TokenDto refreshToken(long userSeq);
 }
