+++ src/main/java/com/example/eshop/member/core/model/UserEntity.java
@@ -1,4 +1,4 @@
-package com.example.eshop.auth.model;
+package com.example.eshop.member.core.model;
 
 import lombok.Builder;
 import lombok.Getter;
