+++ src/main/java/com/example/eshop/member/core/service/MemberService.java
@@ -1,0 +1,10 @@
+package com.example.eshop.member.core.service;
+
+import com.example.eshop.controller.dto.UserDto;
+import com.example.eshop.member.core.model.UserEntity;
+
+public interface MemberService {
+    boolean isDuplicatedId(String id);
+    void signin(UserDto userDto);
+    UserEntity getUserByUserId(String userId);
+}
