+++ src/main/resources/repositories/AuthRepository.xml
@@ -1,57 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.example.eshop.auth.repository.AuthRepository">
-    <select id="isDuplicatedId" parameterType="String" resultType="boolean">
-        SELECT EXISTS(SELECT user_id
-                      FROM tb_member
-                      WHERE 1=1
-                        AND user_id=#{userId})
-    </select>
-
-    <insert id="signin" parameterType="UserEntity">
-        INSERT INTO tb_member (
-        user_id
-        , name
-        , join_cd
-        , status
-        , password
-        , tel
-        , post_num
-        , address
-        , noti_yn
-        , noti_agree_dt
-        , last_login_dt
-        , modify_pw_dt
-        , reg_dt
-        , upd_dt
-        )
-        VALUES (
-        #{userId}
-        , #{name}
-        , #{joinCode}
-        , #{status}
-        , #{password}
-        , #{tel}
-        , #{postNum}
-        , #{address}
-        , #{notiYn}
-        , now()
-        , now()
-        , now()
-        , now()
-        , now()
-        )
-    </insert>
-
-    <select id="findUserByUserId" parameterType="String" resultType="UserEntity">
-        SELECT user_no
-             , user_id
-             , password
-        FROM tb_member
-        WHERE 1=1
-          AND user_id = #{userId}
-    </select>
-
+<mapper namespace="com.example.eshop.member.auth.repository.AuthRepository">
     <select id="findAccessTokenByUserNo" parameterType="long" resultType="TokenEntity">
         SELECT token_no
              , user_no
