+++ src/test/java/com/example/eshop/member/auth/service/impl/AuthServiceImplTest.java
@@ -1,8 +1,9 @@
-package com.example.eshop.auth.service.impl;
+package com.example.eshop.member.auth.service.impl;
 
 import com.example.eshop.controller.dto.LoginDto;
 import com.example.eshop.controller.dto.TokenDto;
 import com.example.eshop.controller.dto.UserDto;
+import com.example.eshop.member.auth.service.impl.AuthServiceImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,29 +19,6 @@ class AuthServiceImplTest {
     @Autowired
     public void setAuthServiceImplTest(AuthServiceImpl authService) {
         this.authService = authService;
-    }
-
-    @Test
-    @DisplayName("isDuplicatedId :: 정상 케이스")
-    void isDuplicatedId() {
-        boolean isDuplicated = authService.isDuplicatedId("hjkim");
-
-        assertTrue(isDuplicated);
-    }
-
-    @Test
-    @Transactional
-    @DisplayName("signin :: 정상 케이스")
-    void signin() {
-        UserDto user = new UserDto("hjkimtest", "test", "01", "test",
-                "01012341234", "000001",
-                "Seoul", "Y");
-
-        authService.signin(user);
-
-        boolean checkRegistered = authService.isDuplicatedId("hjkimtest");
-
-        assertTrue(checkRegistered);
     }
 
     @Test
